refactor: Clean up project structure and consolidate directories
- Move rlhf_core/ into src/rldk/core/ for better organization
- Move utils/ into src/rldk/utils/ to consolidate utility functions
- Remove duplicate train_hf_model.py files and consolidate imports
- Update all import paths to use new consolidated structure
- Fix profiler context usage in training scripts
- Add config/ directory for future configuration files
- Update .gitignore to ignore generated training outputs
- Maintain all existing functionality while improving code organization

This refactoring makes the codebase more navigable and maintainable
without changing any functionality. All training scripts and examples
continue to work as expected.
