+++ .gitignore
@@ -156,3 +156,4 @@ outputs/
 *.lock
 *.zip
 temp_diff/
+hf_training_output/
