+++ tools/run_profile.py
@@ -18,7 +18,7 @@ import torch.optim as optim
 # Add parent directory to path for imports
 sys.path.append(str(Path(__file__).parent.parent))
 
-from rlhf_core.profiler import ProfilerManager
+from rldk.core.profiler import ProfilerManager
 from profiler.torch_profiler import TorchProfiler
 from profiler.profiler_context import ProfilerContext
